# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agrisiti Academy API",
        "description": "REST and WebSocket API for the Agrisiti Academy e-learning platform. Covers student registration, course management, progress tracking, tests, assignments, notes, messaging, community forum, notifications, and the Agri chatbot (AI-powered onboarding assistant with real-time Reverb WebSocket streaming, SSE fallback, and course recommendations). See the Chatbot tag for the complete user journey, quiz questions, and 3-tier transport documentation.",
        "contact": {
            "email": "admin@agrisiti.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://academy-backends.agrisiti.com",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/organisations/register": {
            "post": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Register a new organisation account (pending admin approval before it can post slots)",
                "operationId": "6d85f9f99811db03ceab8d928b4e7a7b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "contact_name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "name"
                                ],
                                "properties": {
                                    "contact_name": {
                                        "description": "Name of the person registering on behalf of the organisation",
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "name": {
                                        "description": "Organisation name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "sector": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "state": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "lga": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "website": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Organisation registered; account is pending admin approval"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organisations/me": {
            "get": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Get the authenticated organisation's own profile",
                "operationId": "a8d054b8107202c4b05f8a0d451ce13b",
                "responses": {
                    "200": {
                        "description": "Organisation profile"
                    },
                    "404": {
                        "description": "No organisation profile for this account"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Update the authenticated organisation's own profile",
                "operationId": "594e8b400b901543dedee41b2e00a1f4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "sector": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "state": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "lga": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "website": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Organisation profile updated"
                    },
                    "404": {
                        "description": "No organisation profile for this account"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Bearer token from /api/login"
            }
        }
    },
...
}